feat: Implemented the wallpaper system for each workspace
fix: Fixed the 'switch_client' feature implementing the configure notify handler

+++ core/workspace/workspace_manager.h
@@ -22,7 +22,7 @@ public:
   void next();
 
 private:
-  int c_workspace = 0;
+  size_t c_workspace = 0;
   std::vector<Workspace> workspaces = {};
   int width, height;
 };
